1. event 分解为 event_handler 及 event 两个类，可以从event.handler() 获取到对应的 event_handler 引用 2. 部分文件编码及统一函数命名［统一utf8编码，获取属性统一用名词格式］

+++ inc/simple/third/url_parser.c
@@ -42,7 +42,7 @@ parse_url(const char *url)
     struct parsed_url *purl;
     const char *tmpstr;
     const char *curstr;
-    int len;
+    long len;
     int i;
     int userpass_flag;
     int bracket_flag;
